refactor(taxonomy): extract DescriptionSection to NodeDescriptionSection (t/2811, t/2812)
Pure move (TL-approved, p/455#73) to stay under the ADR-007 1500-line eslint
max-lines ceiling on NodeDetail.tsx — the ceiling exists to force exactly this,
not to be waived. DescriptionSection + its DescriptionSectionProps/DescriptionMention
interfaces + the local CATEGORY_SINGULAR map move verbatim to a sibling file;
NodeDetail imports DescriptionSection + the DescriptionMention type. Dead imports
(HighlightedTextarea, DescriptionToggle, triggerPovNodeRegeneration, EntityRef,
MentionSegment) removed from NodeDetail. Identical props/exports, zero behavioral
change — this lands t/2811 (plain-desc highlight) + t/2812 (formal-tab auto-switch)
under the gate.

Verify: renderer tsc 0 errors, eslint 0 errors (max-lines cleared; only pre-existing
complexity warnings remain).
