+++ taxonomy-editor/src/renderer/components/taxonomy/NodeDetail.tsx
@@ -9,8 +9,7 @@ import { useTaxonomyStore } from '../../hooks/useTaxonomyStore';
 import type { AggregatedCrux } from '../../hooks/useTaxonomyStore';
 import { useDebateStore } from '../../hooks/useDebateStore';
 import { DeleteConfirmDialog } from '../shared/DeleteConfirmDialog';
-import { HighlightedTextarea } from '../shared/HighlightedField';
-import { buildSearchRegex } from '../../utils/searchRegex';
+import { DescriptionSection, type DescriptionMention } from './NodeDescriptionSection';
 import { TypeaheadSelect } from '../shared/TypeaheadSelect';
 import { FieldHelp } from '../shared/FieldHelp';
 import { TheoryLink } from '../shared/TheoryLink';
@@ -31,12 +30,10 @@ import { nodeTypeFromId, nodePovFromId } from '@lib/debate/nodeIdUtils';
 import { POV_KEYS } from '@lib/debate/types';
 import { api } from '@bridge';
 import { useContainerMentionKit } from '../shared/MentionField';
-import { reconstructNodeContainer, type MentionSegment } from '../shared/mentionText';
-import type { EntityRef } from '@lib/entities/types';
+import { reconstructNodeContainer } from '../shared/mentionText';
 import { EditConflictBadge, type NodeConflict } from '../conflict/edit-conflicts';
-import { triggerPovNodeRegeneration } from '../../utils/regeneratePlainDescription';
 import { generateAphorism } from '../../utils/regenerateAphorism';
-import { useDescriptionMode, resolveDescription, DescriptionToggle } from '../shared/DescriptionToggle';
+import { useDescriptionMode, resolveDescription } from '../shared/DescriptionToggle';
 import { usePreferencesStore } from '../../store/preferencesStore';
 import { EmptyState } from '../shared/EmptyState';
 import { OverflowMenu, type OverflowMenuEntry } from '../shared/OverflowMenu';
@@ -100,11 +97,6 @@ const BDI_GUIDANCE: Record<Category, string> = {
 };
 
 /** Singular form with article for genus-differentia descriptions */
-const CATEGORY_SINGULAR: Record<Category, string> = {
-  'Beliefs': 'A Belief',
-  'Desires': 'A Desire',
-  'Intentions': 'An Intention',
-};
 const POV_LABELS: Record<Pov, string> = {
   accelerationist: POV_META.accelerationist.label,
   safetyist: POV_META.safetyist.label,
@@ -793,95 +785,7 @@ function BeliefsMetricsRow({ node, showDtDrilldown, setShowDtDrilldown }: Belief
   );
 }
 
-// ── Content tab: Description section (extracted from NodeDetail for complexity) ──
-
-/** Entity-mention render segments for the read-only `description` field + their click handler (t/1908). */
-interface DescriptionMention {
-  segments: readonly MentionSegment[];
-  onSelectRef: (ref: EntityRef) => void;
-}
-
-interface DescriptionSectionProps {
-  pov: Pov;
-  node: PovNode;
-  readOnly?: boolean;
-  err: (field: string) => string | undefined;
-  descMode: 'formal' | 'plain';
-  setDescMode: (mode: 'formal' | 'plain') => void;
-  maybeRegenAphorism: () => void;
-  update: (updates: Partial<PovNode>) => void;
-  updatePovNode: (pov: Pov, id: string, updates: Partial<PovNode>) => void;
-  /** Renders a reconstructed container field's text with linkified entity mentions (t/1898); undefined = plain. */
-  renderMentionField?: (fieldName: string, fallback: string) => ReactNode;
-  /** Mention segments for the formal `description` HighlightedTextarea (t/1908); undefined = no links. */
-  descriptionMention?: DescriptionMention;
-  /** t/2811: mention segments for the plain_description read-only highlight path; undefined = no links. */
-  plainDescriptionMention?: DescriptionMention;
-}
-
-function DescriptionSection({ pov, node, readOnly, err, descMode, setDescMode, maybeRegenAphorism, update, updatePovNode, renderMentionField, descriptionMention, plainDescriptionMention }: DescriptionSectionProps) {
-  const viewMode = usePreferencesStore(state => state.viewMode);
-  const { findQuery, findMode, findCaseSensitive } = useTaxonomyStore();
-  // t/2812: a search match in Formal-but-not-Plain surfaces the Formal tab (even under Simple
-  // View); empty query restores the global pref. Fresh regex per test → no /g lastIndex carry.
-  const matchIn = (t: string) => !!findQuery && (buildSearchRegex(findQuery, findMode, findCaseSensitive)?.test(t) ?? false);
-  const baseDescMode: 'formal' | 'plain' = viewMode === 'simple' ? 'plain' : descMode;
-  const effectiveDescMode: 'formal' | 'plain' =
-    baseDescMode === 'plain' && matchIn(node.description ?? '') && !matchIn(node.plain_description ?? '') ? 'formal' : baseDescMode;
-  return (
-    <div className={`form-group ${err('description') ? 'has-error' : ''}`}>
-      <div className="description-header">
-        <label>
-          Description
-          <FieldHelp text={`Genus-differentia format:\n"${CATEGORY_SINGULAR[node.category]} within [POV] discourse that [differentia].\nEncompasses: ...\nExcludes: ..."\nEncompasses and Excludes must each start on a new line.`} />
-        </label>
-        {viewMode === 'advanced' && <DescriptionToggle mode={descMode} onToggle={setDescMode} hasPlainDescription={!!node.plain_description} />}
-      </div>
-      {effectiveDescMode === 'formal' ? (
-        <div className="prose" onBlur={maybeRegenAphorism}>
-          <HighlightedTextarea
-            value={node.description}
-            onChange={(v) => update({ description: v })}
-            rows={6}
-            readOnly={readOnly}
-            mentionSegments={descriptionMention?.segments}
-            onSelectRef={descriptionMention?.onSelectRef}
-          />
-          {err('description') && <div className="error-text">{err('description')}</div>}
-        </div>
-      ) : (
-        <>
-          {node.plain_description === null ? (
-            <div className="plain-description-box plain-description-generating">Regenerating…</div>
-          ) : (
-            <div className="plain-description-box">
-              {/* t/2811: readOnly view highlights search matches + keeps mention links (Formal
-                  already did via HighlightedTextarea). Segments follow the shown field. */}
-              {readOnly ? (
-                <HighlightedTextarea
-                  value={node.plain_description ?? node.description}
-                  readOnly
-                  mentionSegments={(node.plain_description ? plainDescriptionMention : descriptionMention)?.segments}
-                  onSelectRef={(node.plain_description ? plainDescriptionMention : descriptionMention)?.onSelectRef}
-                />
-              ) : (node.plain_description ?? node.description)}
-            </div>
-          )}
-          {!readOnly && (
-            <button
-              type="button"
-              className="plain-description-regen"
-              disabled={node.plain_description === null}
-              onClick={() => triggerPovNodeRegeneration(pov, node.id, node.description, updatePovNode)}
-            >
-              ↻ Regenerate
-            </button>
-          )}
-        </>
-      )}
-    </div>
-  );
-}
+// DescriptionSection extracted to ./NodeDescriptionSection (ADR-007 file-size ceiling; t/2811/t/2812).
 
 // ── Content tab: Steelman Vulnerability section (extracted from NodeDetail for complexity) ──
